rename file for fix case of letters (important for linux)
