+++ src/components/sidebarRight/tabs/gifs.ts
@@ -9,7 +9,7 @@ import InputSearch from "../../inputSearch";
 import animationIntersector from "../../animationIntersector";
 import appSidebarRight from "..";
 import appUsersManager from "../../../lib/appManagers/appUsersManager";
-import appInlineBotsManager, { AppInlineBotsManager } from "../../../lib/appManagers/AppInlineBotsManager";
+import appInlineBotsManager, { AppInlineBotsManager } from "../../../lib/appManagers/appInlineBotsManager";
 import GifsMasonry from "../../gifsMasonry";
 import appImManager from "../../../lib/appManagers/appImManager";
 import type { MyDocument } from "../../../lib/appManagers/appDocsManager";
